requirements/base.txt
Django==1.11.7
django-ckeditor==5.3.1
django-classy-tags==0.8.0
django-cms==3.4.5
django-colorfield==0.1.14
django-filer==1.3.0
django-formtools==2.1
django-js-asset==0.1.1
django-model-utils==3.0.0
django-mptt==0.8.7
django-polymorphic==1.3
django-sekizai==0.10.0
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.4.2
html5lib==0.9999999
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
sorl-thumbnail==12.3
Unidecode==0.4.21
Unipath==1.1
